fix: move ConsoleLogger and it's types to legacy dir
