+++ packages/core/src/logging/index.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export { ConsoleLogger } from './ConsoleLogger';
+export { ConsoleLogger } from './legacy/ConsoleLogger';
 export { LoggingProvider } from './types';
 export { createLogger } from './createLogger';
