+++ packages/core/src/logging/legacy/types.ts
@@ -1,0 +1,40 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Taken from @aws-sdk/client-cloudwatch-logs@3.6.1
+ */
+export interface InputLogEvent {
+	timestamp: number | undefined;
+	message: string | undefined;
+}
+
+export interface LoggingProvider {
+	// return the name of you provider
+	getProviderName(): string;
+
+	// return the name of you category
+	getCategoryName(): string;
+
+	// configure the plugin
+	configure(config?: object): object;
+
+	// take logs and push to provider
+	pushLogs(logs: InputLogEvent[]): void;
+}
+
+export interface Logger {
+	debug(msg: string): void;
+	info(msg: string): void;
+	warn(msg: string): void;
+	error(msg: string): void;
+	addPluggable(pluggable: LoggingProvider): void;
+}
+
+export enum LogType {
+	DEBUG = 'DEBUG',
+	ERROR = 'ERROR',
+	INFO = 'INFO',
+	WARN = 'WARN',
+	VERBOSE = 'VERBOSE',
+}
